Entrega_Semana2/requirements.txt
pygame-ce
esper
pygbag
pyinstaller
